+++ src/PlejdHbPlatform.ts
@@ -8,7 +8,12 @@ import {
   Characteristic,
 } from "homebridge";
 
-import { isAddon, isDimmable, PLATFORM_NAME, PLUGIN_NAME } from "./settings.js";
+import {
+  DEFAULT_BRIGHTNESS_TRANSITION_MS,
+  PLATFORM_NAME,
+  PLUGIN_NAME,
+} from "./constants.js";
+import { isDimmable, isAddon, isSwitch } from "./utils.js";
 import { PlejdHbAccessory } from "./PlejdHbAccessory.js";
 import { UserInputConfig } from "./model/userInputConfig.js";
 import { Device } from "./model/device.js";
@@ -27,6 +32,7 @@ export class PlejdHbPlatform implements DynamicPlatformPlugin {
   public readonly accessories: PlatformAccessory[] = [];
 
   public readonly plejdHbAccessories: PlejdHbAccessory[] = [];
+  private readonly transitionMs: number;
 
   constructor(
     public readonly log: Logger,
@@ -36,6 +42,9 @@ export class PlejdHbPlatform implements DynamicPlatformPlugin {
     homebridgeApi.on("didFinishLaunching", this.configurePlejd);
     this.Characteristic = homebridgeApi.hap.Characteristic;
     this.Service = homebridgeApi.hap.Service;
+
+    this.transitionMs =
+      config.transition_ms ?? DEFAULT_BRIGHTNESS_TRANSITION_MS;
   }
 
   configurePlejd = async () => {
@@ -80,8 +89,10 @@ export class PlejdHbPlatform implements DynamicPlatformPlugin {
         const plejdDevice = site.plejdDevices.find((x) => x.deviceId === id)!;
         const model = plejdDevice.firmware.notes;
 
-        if (isAddon(model)) {
-          this.log.info(`Ignoring ${model} as it is not supported for now`);
+        if (isAddon(model) || isSwitch(model)) {
+          this.log.info(
+            `Ignoring ${model} as it is not supported for now, help with this is needed`,
+          );
           return;
         }
 
@@ -190,7 +201,12 @@ export class PlejdHbPlatform implements DynamicPlatformPlugin {
 
       if (existingAccessory) {
         this.plejdHbAccessories.push(
-          new PlejdHbAccessory(this, existingAccessory, device),
+          new PlejdHbAccessory(
+            this,
+            existingAccessory,
+            device,
+            this.transitionMs,
+          ),
         );
       } else {
         this.addNewDevice(device);
@@ -205,7 +221,9 @@ export class PlejdHbPlatform implements DynamicPlatformPlugin {
     );
     accessory.context.device = device;
     // See above.
-    this.plejdHbAccessories.push(new PlejdHbAccessory(this, accessory, device));
+    this.plejdHbAccessories.push(
+      new PlejdHbAccessory(this, accessory, device, this.transitionMs),
+    );
 
     // link the accessory to your platform
     this.homebridgeApi.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [
@@ -262,7 +280,7 @@ export class PlejdHbPlatform implements DynamicPlatformPlugin {
           ?.updateValue(isOn);
       }
 
-      plejdHbAccessory.updateState(isOn, brightness);
+      plejdHbAccessory.onPlejdUpdates(isOn, brightness);
     } else {
       if (device) {
         this.addNewDevice(device);